AI_Anime_Recommender_using_Grafana_Cloud_Minikube_ChromaDB_Langchain/requirements.txt
langchain-groq
langchain
langchain-community
chromadb
streamlit
pandas
python-dotenv
sentence-transformers
langchain_huggingface